Step 2 - Add Kafka Connect Docker Compose Stack
The bare minimum services in the stack are:
- kafka broker (Kraft enabled)
- schema registry
- kafka connect worker

which are taken from:
- https://github.com/confluentinc/cp-all-in-one/blob/7.9.0-post/cp-all-in-one-community/docker-compose.yml
- The confluentinc/cp-server image is being used with the developer
license to enable metrics and schema registry

Other services from the stack are not included:
- ksqldb-server, ksqldb-cli, ksql-datagen, rest-proxy

The Control Center is taken from:
- https://github.com/confluentinc/cp-all-in-one/blob/7.9.0-post/cp-all-in-one/docker-compose.yml#L81C1-L103C17

The stack also includes a PostgreSQL database and an init script to add
a `pizza_orders` table.

Healthcheck and restart policy have been added to all services.