Step 3.2 - Add Spring Boot Kafka Consumer/JDBC Project
The project is built with gradle and uses Java 21, Spring Kafka, and
Spring JDBC, and kafka-avro-serializer library to consume Avro records.

A simple @KafkaListener is used to consume records, and JdbcOperations
is injected to save the data to the database.

Spring Boot application.yaml has been added with the kafka broker
and JDBC data source configuration.